--- compiler/src/jdk.graal.nativeimage/src/jdk/graal/nativeimage/hosted/GlobalData.java
@@ -1,54 +1,0 @@
-/*
- * Copyright (c) 2025, Oracle and/or its affiliates. All rights reserved.
- * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
- *
- * This code is free software; you can redistribute it and/or modify it
- * under the terms of the GNU General Public License version 2 only, as
- * published by the Free Software Foundation.  Oracle designates this
- * particular file as subject to the "Classpath" exception as provided
- * by Oracle in the LICENSE file that accompanied this code.
- *
- * This code is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
- * version 2 for more details (a copy is included in the LICENSE file that
- * accompanied this code).
- *
- * You should have received a copy of the GNU General Public License version
- * 2 along with this work; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
- *
- * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
- * or visit www.oracle.com if you need additional information or have any
- * questions.
- */
-package jdk.graal.nativeimage.hosted;
-
-import jdk.graal.nativeimage.impl.GlobalDataSupport;
-import org.graalvm.nativeimage.ImageSingletons;
-import org.graalvm.nativeimage.Platform;
-import org.graalvm.nativeimage.Platforms;
-
-import java.util.function.Supplier;
-
-/**
- * Methods for creating initialized, off-heap data that is shared across all isolates in a process.
- *
- * @since 25
- */
-@Platforms(Platform.HOSTED_ONLY.class)
-public final class GlobalData {
-
-    /**
-     * Creates a pre-allocated and pre-initialized word that is off-heap.
-     *
-     * @param initialValue the initial value of the off-heap word
-     * @return a supplier of the address of the off-heap word
-     */
-    public static Supplier<Long> createGlobal(long initialValue) {
-        return ImageSingletons.lookup(GlobalDataSupport.class).createGlobal(initialValue);
-    }
-
-    private GlobalData() {
-    }
-}
