+++ sdk/mx.sdk/mx_sdk.py
@@ -158,7 +158,7 @@ graalvm_sdk_native_image_component = mx_sdk_vm.GraalVmJreComponent(
     third_party_license_files=[],
     dependencies=['sdkc'],
     jar_distributions=[],
-    boot_jars=['sdk:NATIVEIMAGE', 'compiler:GRAAL_NATIVEIMAGE'],
+    boot_jars=['sdk:NATIVEIMAGE', 'sdk:GRAAL_NATIVEIMAGE'],
     stability="supported",
 )
 mx_sdk_vm.register_graalvm_component(graalvm_sdk_native_image_component)
