+++ sdk/src/jdk.graal.nativeimage/snapshot.sigtest
@@ -2,10 +2,7 @@
 #Version
 
 CLSS public abstract interface jdk.graal.nativeimage.LibGraalLoader
-meth public abstract java.nio.file.Path getJavaHome()
-meth public abstract java.util.Set getServicesModules()
-meth public abstract java.lang.ClassLoader getRuntimeClassLoader()
-meth public abstract java.util.Map getModuleMap()
+meth public abstract java.util.Map getClassModuleMap()
 
 CLSS public final jdk.graal.nativeimage.LibGraalRuntime
 meth public static long getIsolateID()
